+++ x509/src/pkix_extensions.rs
@@ -1,6 +1,6 @@
 //! Extensions [`Extensions`] as defined in RFC 5280
 
-use crate::ext::pkix::name::{DistributionPointName, GeneralName, GeneralNames};
+use crate::ext::pkix::name::{DistributionPointName, GeneralNames};
 
 use alloc::vec::Vec;
 
@@ -41,44 +41,6 @@ pub struct PrivateKeyUsagePeriod {
     #[asn1(context_specific = "1", tag_mode = "IMPLICIT", optional = "true")]
     pub not_after: Option<GeneralizedTime>,
 }
-
-/// Authority information access extension as defined in [RFC 5280 Section 4.2.2.1] and as identified by the [`PKIX_PE_AUTHORITYINFOACCESS`](constant.PKIX_PE_AUTHORITYINFOACCESS.html) OID.
-///
-/// ```text
-/// AuthorityInfoAccessSyntax  ::=
-///         SEQUENCE SIZE (1..MAX) OF AccessDescription
-/// ```
-///
-/// [RFC 5280 Section 4.2.2.1]: https://datatracker.ietf.org/doc/html/rfc5280#section-4.2.2.1
-pub type AuthorityInfoAccessSyntax<'a> = Vec<AccessDescription<'a>>;
-
-/// AccessDescription as defined in [RFC 5280 Section 4.2.2.1].
-///
-/// ```text
-/// AccessDescription  ::=  SEQUENCE {
-///     accessMethod          OBJECT IDENTIFIER,
-///     accessLocation        GeneralName
-/// }
-/// ```
-///
-/// [RFC 5280 Section 4.2.2.1]: https://datatracker.ietf.org/doc/html/rfc5280#section-4.2.2.1
-#[derive(Clone, Debug, Eq, PartialEq, Sequence)]
-#[allow(missing_docs)]
-pub struct AccessDescription<'a> {
-    pub access_method: ObjectIdentifier,
-
-    pub access_location: GeneralName<'a>,
-}
-
-/// Subject information access extension as defined in [RFC 5280 Section 4.2.2.2] and as identified by the [`PKIX_PE_SUBJECTINFOACCESS`](constant.PKIX_PE_SUBJECTINFOACCESS.html) OID.
-///
-/// ```text
-/// SubjectInfoAccessSyntax  ::=
-///         SEQUENCE SIZE (1..MAX) OF AccessDescription
-/// ```
-///
-/// [RFC 5280 Section 4.2.2.2]: https://datatracker.ietf.org/doc/html/rfc5280#section-4.2.2.2
-pub type SubjectInfoAccessSyntax<'a> = Vec<AccessDescription<'a>>;
 
 /// CRL number extension as defined in [RFC 5280 Section 5.2.3] and as identified by the [`PKIX_CE_CRLNUMBER`](constant.PKIX_CE_CRLNUMBER.html) OID.
 ///
